requirements/development.txt
-r base.txt

# Base
# ------------------------------------------------------------------------------
invoke
rich
ipdb==0.13.9
Werkzeug==2.0.3
psycopg2-binary==2.9.3
watchgod

# Testing
# ------------------------------------------------------------------------------
django-stubs==1.9.0
pytest==7.1.2
pytest-sugar==0.9.4
pytest-django==4.5.2
djangorestframework-stubs==1.4.0

# Code quality
# ------------------------------------------------------------------------------
flake8==4.0.1
autopep8
flake8-isort==4.1.1
coverage==6.4.1
black==22.3.0
pre-commit

# Django
# ------------------------------------------------------------------------------
factory-boy==3.2.1
django-extensions==3.1.5
django-coverage-plugin==2.0.3
django-debug-toolbar==3.4.0 
